refactor(team): regroup pane-dispatch infra + sweep thin wrappers (C15)
Two semantic improvements bundled into one PR:

1. **Move pane-dispatch infrastructure to pane_dispatch.go** —
   paneDispatchTurn (the on-the-wire turn type),
   paneDispatchMinGap + paneDispatchCoalesceWindow (dispatcher
   knobs), launcherSendNotificationToPaneFn (the test-seam fn type),
   launcherSendNotificationToPaneOverride (the atomic.Pointer seam),
   and launcherSendNotificationToPane (the seam-aware send entry
   point) all lived in launcher.go. They belong next to the
   dispatcher type they support, not in the orchestrator file.

2. **Sweep thin pane / scheduler wrappers (PLAN.md §6)**.
   Like C9 did for the targeter wrappers, this PR deletes the
   one-line Launcher methods that just delegated to a sub-type.
   Call sites updated via sed:

     l.channelPaneStatus()         -> l.panes().ChannelPaneStatus()
     l.captureDeadChannelPane(s)   -> l.panes().CaptureDeadChannelPane(s)
     l.clearAgentPanes()           -> l.panes().ClearAgentPanes()
     l.clearOverflowAgentWindows() -> l.panes().ClearOverflowAgentWindows()
     l.listTeamPanes()             -> l.panes().ListTeamPanes()
     l.spawnVisibleAgents()        -> l.panes().SpawnVisibleAgents()
     l.spawnOverflowAgents()       -> l.panes().SpawnOverflowAgents()
     l.detectDeadPanesAfterSpawn(m)-> l.panes().DetectDeadPanesAfterSpawn(m)
     l.capturePaneTargetContent(t) -> l.panes().CapturePaneTargetContent(t)
     l.capturePaneContent(idx)     -> l.panes().CapturePaneContent(idx)

     l.processDueSchedulerJobs()   -> l.scheduler().processOnce()
     l.processDueTaskJob(j)        -> l.scheduler().processTaskJob(j)
     l.processDueRequestJob(j)     -> l.scheduler().processRequestJob(j)
     l.processDueWorkflowJob(j)    -> l.scheduler().processWorkflowJob(j)
     l.recordWatchdogLedger(...)   -> l.scheduler().recordLedger(...)

   trySpawnWebAgentPanes / reportPaneFallback wrappers deleted
   without renaming because they had zero live call sites
   (referenced only in doc comments).

launcher.go: 865 -> 786 lines (-79).

Cumulative since main: 4998 -> 786 = -4212 lines (-84.3%).

Stacked on refactor/launcher-c14-manifest-membership (C14).
